% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_Convert_Plot.R
\name{tsConvertPlot}
\alias{tsConvertPlot}
\title{tsConvertPlot}
\usage{
tsConvertPlot(data)
}
\arguments{
\item{data}{Data you want to transform to time series and be plotted}
}
\value{

}
\description{
tsConvertPlot
}
\examples{
data = rnorm(30,10,2)
ts_plot(data)
}
